Move backtrace to the HAL folder (#10790)
