+++ Marlin/src/HAL/HAL_DUE/DebugMonitor_Due.cpp
@@ -22,10 +22,12 @@
 
 #ifdef ARDUINO_ARCH_SAM
 
-#include "../../inc/MarlinConfig.h"
-#include "../../Marlin.h"
-#include "../../backtrace/unwinder.h"
-#include "../../backtrace/unwmemaccess.h"
+#include "../../core/macros.h"
+#include "../../core/serial.h"
+#include <stdarg.h>
+
+#include "../backtrace/unwinder.h"
+#include "../backtrace/unwmemaccess.h"
 
 // Debug monitor that dumps to the Programming port all status when
 // an exception or WDT timeout happens - And then resets the board
@@ -65,7 +67,6 @@ static void TXBegin(void) {
   // Enable receiver and transmitter
   UART->UART_CR = UART_CR_RXEN | UART_CR_TXEN;
 }
-
 
 // Send character through UART with no interrupts
 static void TX(char c) {
@@ -333,4 +334,4 @@ __attribute__((naked)) void RSTC_Handler(void) {
   );
 }
 
-#endif
+#endif // ARDUINO_ARCH_SAM
